refactored summary list
