+++ src/main/java/seedu/planner/logic/commands/SummaryCommand.java
@@ -17,5 +17,6 @@ public abstract class SummaryCommand extends Command {
             + "Example: \"" + COMMAND_WORD + " date " + PREFIX_DATE + "18-9-2018 " + "20-9-2018\"\n"
             + "OR \"" + COMMAND_WORD + " month " + PREFIX_DATE + "apr-2018 " + "may-2018\"\n"
             + "OR \"" + COMMAND_WORD + " category " + PREFIX_DATE + "18-9-2018 " + "20-9-2018\"";
+
     public static final String TOTAL_LABEL = "TOTAL";
 }
