+++ src/main/java/seedu/planner/model/summary/SummaryList.java
@@ -1,6 +1,7 @@
 package seedu.planner.model.summary;
 
 import javafx.collections.ObservableList;
+import seedu.planner.commons.util.MoneyUtil;
 import seedu.planner.model.record.MoneyFlow;
 import seedu.planner.model.record.Record;
 import seedu.planner.ui.SummaryEntry;
@@ -11,6 +12,16 @@ import seedu.planner.ui.SummaryEntry;
  * and implements only list functions
  */
 public abstract class SummaryList {
+
+    protected MoneyFlow total;
+    protected MoneyFlow totalIncome;
+    protected MoneyFlow totalExpense;
+
+    public SummaryList() {
+        total = new MoneyFlow("-0");
+        totalIncome = new MoneyFlow("-0");
+        totalExpense = new MoneyFlow("-0");
+    }
 
     /**
      * Returns the simple class name of the identifier used to categorise each entry in list
@@ -48,5 +59,18 @@ public abstract class SummaryList {
      */
     public abstract ObservableList<SummaryEntry> getSummaryList();
 
+    /** Update the total moneyflow, total income and total expense */
+    protected void updateTotals(Record record) {
+        MoneyFlow money = record.getMoneyFlow();
+        if (isExpense(money)) {
+            totalExpense = MoneyUtil.add(totalExpense, money);
+        } else {
+            totalIncome = MoneyUtil.add(totalIncome, money);
+        }
+        total = MoneyUtil.add(total, money);
+    }
 
+    private boolean isExpense(MoneyFlow money) {
+        return money.toDouble() < 0;
+    }
 }
